+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.6 (WIP)
+## 0.2.6 (2023-06-17)
 
 * Fix `al-field-schema` element not being able to get schema of an arrayable field
 * Add `Pathway` system for faceted breadcrumbs